refactor: move graph_sync_metrics into framework
Shared graph sync sampling helpers belong with framework utilities, not under test_cases/fiber.
